refactor(portfolio): extract code block action buttons as slot
